edge-device/requirements.txt
# Audio Processing
pyaudio>=0.2.13
numpy>=1.21.0
soundfile>=0.10.3

# GPS
gpsd-py3>=2.3.0

# File Operations
python-dotenv>=0.19.0
watchdog>=2.1.6

# Networking
requests>=2.26.0
paramiko>=2.8.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
